+++ packages/headless/src/controllers/core/facets/facet/headless-core-facet.ts
@@ -18,7 +18,10 @@ import {
   updateFacetNumberOfValues,
   updateFacetIsFieldExpanded,
 } from '../../../../features/facets/facet-set/facet-set-actions';
-import {executeToggleFacetSelect} from '../../../../features/facets/facet-set/facet-set-controller-actions';
+import {
+  executeToggleFacetExclude,
+  executeToggleFacetSelect,
+} from '../../../../features/facets/facet-set/facet-set-controller-actions';
 import {
   facetRequestSelector,
   facetResponseSelector,
@@ -26,7 +29,10 @@ import {
 } from '../../../../features/facets/facet-set/facet-set-selectors';
 import {facetSetReducer as facetSet} from '../../../../features/facets/facet-set/facet-set-slice';
 import {defaultFacetOptions} from '../../../../features/facets/facet-set/facet-set-slice';
-import {isFacetValueSelected} from '../../../../features/facets/facet-set/facet-set-utils';
+import {
+  isFacetValueExcluded,
+  isFacetValueSelected,
+} from '../../../../features/facets/facet-set/facet-set-utils';
 import {FacetSortCriterion} from '../../../../features/facets/facet-set/interfaces/request';
 import {
   ConfigurationSection,
@@ -89,11 +95,25 @@ export interface CoreFacet extends Controller {
   toggleSelect(selection: FacetValue): void;
 
   /**
+   * Toggles exclusion of the specified facet value.
+   *
+   * @param selection - The facet value to toggle exclusion.
+   */
+  toggleExclude(selection: FacetValue): void;
+
+  /**
    * Toggles the specified facet value, deselecting others.
    *
    * @param selection - The facet value to toggle.
    */
   toggleSingleSelect(selection: FacetValue): void;
+
+  /**
+   * Excludes the specified facet value, deselecting others.
+   *
+   * @param selection - The facet value to toggle exclusion.
+   */
+  toggleSingleExclude(selection: FacetValue): void;
 
   /**
    * Checks whether the specified facet value is selected.
@@ -102,6 +122,14 @@ export interface CoreFacet extends Controller {
    * @returns Whether the specified facet value is selected.
    */
   isValueSelected(value: FacetValue): boolean;
+
+  /**
+   * Checks whether the specified facet value is excluded.
+   *
+   * @param value - The facet value to check.
+   * @returns Whether the specified facet value is excluded.
+   */
+  isValueExcluded(value: FacetValue): boolean;
 
   /**
    * Deselects all facet values.
@@ -209,11 +237,25 @@ export interface FacetSearch {
   select(value: SpecificFacetSearchResult): void;
 
   /**
+   * Excludes a facet search result.
+   *
+   * @param value - The search result to exclude.
+   * */
+  exclude(value: SpecificFacetSearchResult): void;
+
+  /**
    * Selects a search result while deselecting facet values.
    *
    * @param value - The search result to select.
    * */
   singleSelect(value: SpecificFacetSearchResult): void;
+
+  /**
+   * Excludes a search result while including facet values.
+   *
+   * @param value - The search result to exclude.
+   * */
+  singleExclude(value: SpecificFacetSearchResult): void;
 
   /**
    * Resets the query and empties the values.
@@ -342,6 +384,11 @@ export function buildCoreFacet(
     toggleSelect: (selection: FacetValue) =>
       dispatch(executeToggleFacetSelect({facetId: options.facetId, selection})),
 
+    toggleExclude: (selection: FacetValue) =>
+      dispatch(
+        executeToggleFacetExclude({facetId: options.facetId, selection})
+      ),
+
     // Must use a function here to properly support inheritance with `this`.
     toggleSingleSelect: function (selection: FacetValue) {
       if (selection.state === 'idle') {
@@ -351,16 +398,27 @@ export function buildCoreFacet(
       this.toggleSelect(selection);
     },
 
+    // Must use a function here to properly support inheritance with `this`.
+    toggleSingleExclude: function (selection: FacetValue) {
+      if (selection.state === 'idle') {
+        dispatch(deselectAllFacetValues(facetId));
+      }
+
+      this.toggleExclude(selection);
+    },
+
     isValueSelected: isFacetValueSelected,
+
+    isValueExcluded: isFacetValueExcluded,
 
     deselectAll() {
       dispatch(deselectAllFacetValues(facetId));
-      dispatch(updateFacetOptions({freezeFacetOrder: true}));
+      dispatch(updateFacetOptions());
     },
 
     sortBy(criterion: FacetSortCriterion) {
       dispatch(updateFacetSortCriterion({facetId, criterion}));
-      dispatch(updateFacetOptions({freezeFacetOrder: true}));
+      dispatch(updateFacetOptions());
     },
 
     isSortedBy(criterion: FacetSortCriterion) {
@@ -376,7 +434,7 @@ export function buildCoreFacet(
 
       dispatch(updateFacetNumberOfValues({facetId, numberOfValues}));
       dispatch(updateFacetIsFieldExpanded({facetId, isFieldExpanded: true}));
-      dispatch(updateFacetOptions({freezeFacetOrder: true}));
+      dispatch(updateFacetOptions());
     },
 
     showLessValues() {
@@ -390,7 +448,7 @@ export function buildCoreFacet(
         updateFacetNumberOfValues({facetId, numberOfValues: newNumberOfValues})
       );
       dispatch(updateFacetIsFieldExpanded({facetId, isFieldExpanded: false}));
-      dispatch(updateFacetOptions({freezeFacetOrder: true}));
+      dispatch(updateFacetOptions());
     },
 
     enable() {