+++ packages/headless/src/controllers/recommendation/headless-recommendation.test.ts
@@ -1,4 +1,4 @@
-import {Action} from 'redux';
+import {Action} from '@reduxjs/toolkit';
 import {configuration} from '../../app/common-reducers';
 import {updateNumberOfResults} from '../../features/pagination/pagination-actions';
 import {